api_get_order_by_id

+++ DataAccess/DAO/OrderDAO.cs
@@ -92,7 +92,7 @@ public class OrderDAO : IBaseDAO<Order>
     {
         try
         {
-            return await _context.Orders.FindAsync(id);
+            return await _context.Orders.FindAsync((int)id);
         }
         catch (Exception e)
         {
